More code sharing to satisfy lint dup code detector
