+++ sdk/python/lib/pulumi/automation/events.py
@@ -17,6 +17,7 @@
 
 from enum import Enum
 from typing import Optional, List, Mapping, Any, MutableMapping
+from ._representable import _Representable
 
 
 class OpType(str, Enum):
@@ -43,13 +44,8 @@ class OpType(str, Enum):
 OpMap = MutableMapping[OpType, int]
 
 
-class BaseEvent:
-    def __repr__(self):
-        # pylint: disable=duplicate-code
-        inputs = self.__dict__
-        fields = [f"{key}={inputs[key]!r}" for key in inputs]  # pylint: disable=consider-using-dict-items
-        fields = ", ".join(fields)
-        return f"{self.__class__.__name__}({fields})"
+class BaseEvent(_Representable):
+    pass
 
 
 class CancelEvent(BaseEvent):
